+++ enterprise/api/v1/views.py
@@ -15,6 +15,7 @@ from rest_framework import filters, generics, permissions, status, viewsets
 from rest_framework.authentication import SessionAuthentication
 from rest_framework.decorators import action
 from rest_framework.exceptions import NotFound, ValidationError
+from rest_framework.mixins import CreateModelMixin
 from rest_framework.pagination import PageNumberPagination
 from rest_framework.parsers import FormParser, MultiPartParser
 from rest_framework.permissions import IsAuthenticated
@@ -151,6 +152,14 @@ class EnterpriseReadWriteModelViewSet(EnterpriseModelViewSet, viewsets.ModelView
     permission_classes = (permissions.IsAuthenticated, permissions.DjangoModelPermissions,)
 
 
+class EnterpriseWriteOnlyModelViewSet(EnterpriseModelViewSet, CreateModelMixin, viewsets.GenericViewSet):
+    """
+    Base class for all write only Enterprise model view sets.
+    """
+
+    permission_classes = (permissions.IsAuthenticated, permissions.DjangoModelPermissions)
+
+
 class EnterpriseCustomerViewSet(EnterpriseReadWriteModelViewSet):
     """
     API views for the ``enterprise-customer`` API endpoint.
@@ -184,14 +193,19 @@ class EnterpriseCustomerViewSet(EnterpriseReadWriteModelViewSet):
     def basic_list(self, request, *arg, **kwargs):
         """
         Enterprise Customer's Basic data list without pagination
+
+        Two query parameters are supported:
+        - name_or_uuid: filter by name or uuid substring search in a single query parameter.
+        Primarily used for frontend debounced input search.
+        - startswith: filter by name starting with the given string
         """
         startswith = request.GET.get('startswith')
-        q = request.GET.get('q')
+        name_or_uuid = request.GET.get('name_or_uuid')
         queryset = self.get_queryset().order_by('name')
         if startswith:
             queryset = queryset.filter(name__istartswith=startswith)
-        if q:
-            queryset = queryset.filter(Q(name__icontains=q) | Q(uuid__icontains=q))
+        if name_or_uuid:
+            queryset = queryset.filter(Q(name__icontains=name_or_uuid) | Q(uuid__icontains=name_or_uuid))
         serializer = self.get_serializer(queryset, many=True)
         return Response(serializer.data)
 
@@ -1210,6 +1224,12 @@ class EnterpriseCustomerBrandingConfigurationViewSet(EnterpriseReadWriteModelVie
             )
             return Response("Error with updating branding configuration", status=status.HTTP_500_INTERNAL_SERVER_ERROR)
         return Response("Branding was updated", status=status.HTTP_204_NO_CONTENT)
+
+
+class EnterpriseCustomerCatalogWriteViewSet(EnterpriseWriteOnlyModelViewSet):
+    queryset = models.EnterpriseCustomerCatalog.objects.all()
+    permission_classes = (permissions.IsAdminUser,)
+    serializer_class = serializers.EnterpriseCustomerCatalogWriteOnlySerializer
 
 
 class EnterpriseCustomerCatalogViewSet(EnterpriseReadOnlyModelViewSet):